Remove git dependency from setup.sh
In order for us to have this script, git already has to be installed
on the system. Remove it as a dependency in the `apt-get install`
line of the setup.sh.

Alternative to consider: Instructions in the book could use a wget
command to pull in the raw setup.sh file, so we don't need to instruct
the user to install git.